Made minor changes to all java files
